Refactors code
- Modularized code. Arguments parsing happen in args.rs
- Main is only responsible for showing error messages and calling radeco
